pipeline-automation-scripts/sqoop_job/requirements.txt
grpcio==1.33.2
google-auth==1.23.0
google-cloud==0.34.0
google-cloud-storage==1.31.2
google-cloud-dataproc==2.0.2
